refactor(metrics): unify primitive import path and decorator source (#472)
Rename the pipeline primitive modules from compute_<x>.py to _<x>.py so a
module no longer shares the name of the function it defines. The package
re-export `from factrix.metrics._primitives import compute_<x>` is now
unambiguous, so every consumer (library and tests) uses that single path
and mypy can resolve the symbols — replacing the scattered submodule
imports added as a workaround.

Import the `metric` decorator from factrix.metrics._decorators everywhere
instead of mixing the package surface and the submodule, keeping one
canonical source and avoiding the package import cycle.

Refs #472
