+++ factrix/metrics/_primitives/__init__.py
@@ -1,11 +1,11 @@
-from factrix.metrics._primitives.compute_caar import compute_caar
-from factrix.metrics._primitives.compute_event_returns import compute_event_returns
-from factrix.metrics._primitives.compute_fm_betas import compute_fm_betas
-from factrix.metrics._primitives.compute_group_returns import compute_group_returns
-from factrix.metrics._primitives.compute_ic import compute_ic
-from factrix.metrics._primitives.compute_mfe_mae import compute_mfe_mae
-from factrix.metrics._primitives.compute_spread_series import compute_spread_series
-from factrix.metrics._primitives.compute_ts_betas import compute_ts_betas
+from factrix.metrics._primitives._caar import compute_caar
+from factrix.metrics._primitives._event_returns import compute_event_returns
+from factrix.metrics._primitives._fm_betas import compute_fm_betas
+from factrix.metrics._primitives._group_returns import compute_group_returns
+from factrix.metrics._primitives._ic import compute_ic
+from factrix.metrics._primitives._mfe_mae import compute_mfe_mae
+from factrix.metrics._primitives._spread_series import compute_spread_series
+from factrix.metrics._primitives._ts_betas import compute_ts_betas
 
 __all__ = [
     "compute_caar",
