+++ factrix/metrics/_primitives/_caar.py
