+++ factrix/metrics/_primitives/_group_returns.py
