+++ factrix/metrics/_primitives/_spread_series.py
