+++ factrix/metrics/concentration.py
@@ -38,7 +38,7 @@ from factrix._types import (
     MIN_PORTFOLIO_PERIODS_WARN,
     ConcentrationWeight,
 )
-from factrix.metrics import metric
+from factrix.metrics._decorators import metric
 from factrix.metrics._helpers import (
     _compute_tie_ratio,
     _sample_non_overlapping,
