+++ factrix/metrics/corrado_rank.py
@@ -21,7 +21,7 @@ from factrix._metric_index import cell
 from factrix._results import MetricResult
 from factrix._stats import _calc_t_stat, _p_value_from_z
 from factrix._types import EPSILON, MIN_EVENTS_HARD
-from factrix.metrics import metric
+from factrix.metrics._decorators import metric
 from factrix.metrics._helpers import _short_circuit_output
 
 __all__ = [
