+++ factrix/metrics/ts_beta.py
@@ -36,9 +36,9 @@ from factrix._stats import (
     _p_value_from_t,
 )
 from factrix._types import DDOF
-from factrix.metrics import metric
+from factrix.metrics._decorators import metric
 from factrix.metrics._helpers import _short_circuit_output
-from factrix.metrics._primitives.compute_ts_betas import compute_ts_betas
+from factrix.metrics._primitives import compute_ts_betas
 
 __all__ = [  # noqa: RUF022 (teaching order, see #322 SSOT note)
     "ts_beta",
